+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 
 ### BREAKING CHANGES
 
-- **viewport:** `ViewportSizeType` has been renamed to `UxViewportSizes` and is now a dictionary.
+- **viewport:** `ViewportSizeType` is now a dictionary.
 
 ## [1.1.1](https://github.com/sketch7/ngx.ux/compare/1.1.0...1.1.1) (2020-05-29)